Created the fsdd (File system disk driver) continued moving all disk related logic to it.
fsdd calls read and write methods in disk.js representing the actual
read and write operations of the disk.

lots of special getter / settor methods created in fsdd to manage the
creation of blocks in both the directory and data blocks and access to
pieces of information in the blocks.